# docs/openapi/recovery-v1.yaml
openapi: 3.0.3
info:
  title: InvoNotify Recovery API
  version: 1.0.0
  description: |
    Tenant-scoped recovery APIs for the AI Revenue Recovery demo. All write
    endpoints require an authenticated merchant session and reject cross-origin
    state changes. Recovery action execution is simulation-only by default.
servers:
  - url: http://localhost:3001
security:
  - sessionCookie: []
paths:
...
  /api/v1/recovery-cases/{caseId}/promises:
    get:
      summary: Fetch the active promise-to-pay for a case
      parameters: [{ $ref: '#/components/parameters/caseId' }]
      responses: { '200': { description: Promise record } }
    post:
      summary: Capture a validated promise-to-pay
      parameters: [{ $ref: '#/components/parameters/caseId' }]
      responses: { '201': { description: Created promise } }
...
components:
  securitySchemes:
    sessionCookie: { type: apiKey, in: cookie, name: authjs.session-token }
  parameters:
...
    caseId: { name: caseId, in: path, required: true, schema: { type: integer, minimum: 1 } }
...